style: Created File for RoleCommands
It makes more sense to put the Role Commands in an extra File
