+++ Program.cs
@@ -12,7 +12,6 @@ using DSharpPlus.CommandsNext;
 using MySqlX.XDevAPI;
 using Client = MySql.Data.MySqlClient.Memcached.Client;
 using DSharpPlus.EventArgs;
-using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -138,43 +137,5 @@ namespace DiscordBot
         }
   }
 
-  public class RoleCommands : BaseCommandModule
-  { 
-      
-      [Command("saverolewithpw")]
-      public async Task AddKeyCodeToRole(CommandContext ctx, string rolename, string keycode)
-      {
-          await SaveRoleToDB(rolename, keycode, ctx);
-      }
-      [Command("updaterolewithpw")]
-      public async Task UpdateKeyCodeFromRole(CommandContext ctx, string rolename, string keycode)
-      {
-          await UpdateRoleInDB(rolename, keycode, ctx);
-      }
-
-      private async Task SaveRoleToDB(string customrolename, string customkeycode, CommandContext ctx)
-      {
-          var db = new Database();
-          //string sqlQuery = $"INSERT INTO CustomRoles (rolename, keycode) VALUES ('{customrolename}', '{customkeycode}')";
-          string sqlQuery = $"SELECT * FROM CustomRoles WHERE rolename='{customrolename}'";
-          string returnsql = db.runSQL(sqlQuery);
-          if (returnsql == "1")
-          {
-              await ctx.RespondAsync("This Role already exists with a Password. If you want to update the Password use !updaterolewithpw {rolename} {keycode}");
-          }
-          else
-          {
-              sqlQuery = $"INSERT INTO CustomRoles (rolename, keycode) VALUES ('{customrolename}', '{customkeycode}')";
-              db.runSQL(sqlQuery);
-          }
-      }
-
-      private async Task UpdateRoleInDB(string customrolename, string customkeycode, CommandContext ctx)
-      {
-          var db = new Database();
-
-          string sqlQuery = $"UPDATE CustomRoles SET keycode = '{customkeycode}' WHERE rolename='{customrolename}'";
-          db.runSQL(sqlQuery);
-      }
-    }
+  
 }
