--- Rolemanager/RoleKeyCode.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DiscordBot.Rolemanager
-{
-    class RoleKeyCode
-    {
-        public string Role { get; set; }
-        public string KeyCode { get; set; }
-
-
-    }
-}
