--- Rolemanager/role-config.json
@@ -1,1 +1,0 @@
-"{Testff : TEst123}"
